refactoring: move server and client logic into separate files
